splitting the code in half
